+++ R/ChromBackendMemory-functions.R
@@ -1,7 +1,0 @@
-#' @rdname ChromBackendMemory
-#'
-#' @importFrom methods new
-#' @export ChromBackendMemory
-ChromBackendMemory <- function() {
-    new("ChromBackendMemory")
-}