+++ tests/testthat/test_ChromBackend.R
@@ -57,6 +57,18 @@ test_that("ChromBackend methods throw errors", {
     expect_error(dm[["character", "character2"]], "is not supported")
     expect_error(dm[[123]] <- 123, "is supposed to be a character")
     expect_error(dm[["character", "character2"]] <- 123, "is not supported")
+    expect_identical(filterPeaksData(dm, ranges = c()), dm)
+    expect_identical(filterPeaksData(dm, ranges = c(1,2), variables = c()), dm)
+    expect_error(filterPeaksData(dm, ranges = c("a","3"), variables = c("a")),
+                 "only support")
+    expect_error(filterPeaksData(dm, ranges = c(1,2), variables = c("a")),
+                 "One or more")
+    expect_error(filterPeaksData(dm, ranges = c(1,2), variables = c(33)),
+                 "needs to be of type")
+    expect_error(filterPeaksData(dm, ranges = c(1, 2),
+                                 variables = c("intensity", "rtime")),
+                 "needs to be twice")
+
 })